requirements.txt
aiogram==2.25.2
tinydb
aiohttp
beautifulsoup4
Pillow
qrcode
pyzbar
pathlib